Basics of Haskell "schema"
This provides a basic mirroring of our schema, in Haskell code. Each entity is
represented as a data type with record syntax. Fields that may be null use the
Maybe kind in order to specify if they are null or have a value.

For foreign keys, the situation is a little more intrict. When we load an
edition, for example, all we know about the book is its BookId. In this case, we
can store that information in a Book, with the BookReference constructor.
We use Just BookReference, because we know there is *some* book out
there. However, attempts to get the book name will fail, because bookName
requires something constructed with Book, not BookReference.