Introduce the copointed WithGid type, to keep track of entities that have a GID
When we come to create entities later, we want to automatically generate a GID
for them, which means we won't have anything to stick in for `personGid' for
example. WithGid simply wraps a data-type and ensure that it has a GID as well.

There is one problem now, where accessing data for a `WithGid a' type requires
an extra level of dereferencing; eg personName $ copoint person. I think we can
get around this by using functions like `personName :: Copointed a => a -> Text'
and making use of the Identity type (in the case where we don't know the GID,
such as creating a new Person). It does through one spanner in the work wrt
record constructors though, and hence I have held off on this for now.