Introduce BrainzStem 'LoadedEntity' and 'LoadedCoreEntity' types
A core entity is an entity that has a GID, and an ID in the database. An entity
is something that has just an ID in the database.

By factoring the storage of these rows out, we now have the possibility to pass
around 'Book' objects for things like inserting into the database, previewing,
and much more.

I'm not set on the separation of LoadedCoreEntity and LoadedEntity, but at least
the InDatabase type class allows these to be mostly used transparently.