Use a versioned SQL schema
This change splits the main schema into 2 parts: bookbrainz, and
bookbrainz_v. bookbrainz_v is a schema which keeps track of data that is
versioned, that is - data with a history. This works by introducing a few more
concepts...

First we add the notion of a 'version' to each of the core entities. We then add
'revision' and 'branch' objects, giving each object a 'master' branch by
default. We add views back to the original schema, using the latest revision in
the master branch of each object.