Drastically refactor code to use Snaplets (Snap 0.6)
SnapExtend has been replaced with Snaplets, so I've moved all the code over to
use them. In the process, I've taken the liberty to drastically move bits and
pieces around. All of the web specific functionality now lives in the
BookBrainz.Web namespace, and non-web (and non-Snap) specific functionality is
now in the general BookBrainz namespace. The plan for this separation is to make
writing scripts easy, and not depending on the whole Snap stack.