Drastically document and move code around
I hope this is the large big change I need to make to the codebase for a while
now. We are using Snap 0.6 Snaplets for our infrastructure, digestive-functors
for forms, web-routes-boomerang for routing and Blaze for HTML generation. All
web specific code is now in BookBrainz.Web, the rest is simply under the
BookBrainz namespace.