Adding MusicBrainz like styling
pageLayout has been extended to now take a (Maybe Html) parameter
first. This is the content of the sidebar. A new component,
detailTable has been added as well, which can be used to create full
width tables with alternatiting row formatting.

style.css has been added which is a simplification of the current
MusicBrainz site.