Abstract the versioning over to BrainzStem
This adds the 'insert' and 'update' functions to the 'CoreEntity' type
class. These have default implementations, but do require the 'addVersion'
function to be implemented. This is currently doing a bit more than it needs to
(it manages the revision as well), and can be stripped down a little more.

Currently, only Book has addVersion defined.