Make use of the new entity level references
This schema change makes a few changes to cross-referencing entities.

* book_person_role now references a general person, and a specific book
  commit. This means that a commit can change the person that a role refers to.
* edition_person_role now references a general person, and a specific edition
  commit. Same reason as above.
* edition_revisions now generally references a publisher. This allows the
  publisher to change over specific revisions.
* edition generally refers to a book. This means that over all, an edition
  refers to a book.

A few things to note. The book reference from edition means that an edition
cannot be moved via a commit. This will probably change, which leads on to the
next point...

I don't think commit is the right place to keep track of cross referencing
data. In Git, this would be a tree object, and I think we might need to
introduce tree objects soon here too (yet another abstraction...). In this
model, a book tree would contain the book version and book_person_roles. Edition
tree's would contain a publisher reference, an edition version and
edition_person_roles.