Introduce _tree tables
The _tree tables provide a way to link various bits of data into a single
cohesive entity. In the case of a book, this is the version and book-person
roles. In the case of an edition, it's the book the edition belongs to, the
publisher who published it, the version of the edition itself, and all
edition-person roles.

In 94f8f238af2472635a8865e499bfbbc40e9cff3d I mentioned that I wasn't happy
storing these links in the commit object, so this simply expands on that line of
thought introducing the tree objects.