Rethink the model API
I was unhappy with the definition of the BrainzStem.Model type clases, and have
redesigned them. We now have a nice set of building blocks that work with the
versioning system, and some generic abstractions that build on top of them. That
lets us have a single definition for 'create', 'update', and friends. The
essential buliding blocks that vary from type to type are defined in the
CoreEntity type class.

As most CoreEntity's at the moment have an identical implementation, I've
introduced the GenericVersioning type class, which provides an instance with the
minimal amount of boiler plate. This is now one bit of code I'm still unhappy
with, because it requires UndecidableInstances. This hasn't been a problem yet,
but I bet it could be in the future. Luckily, any problems will be at compile
time I believe, so we'll certainly know when we hit them.