Introduce the BBID table and s/gid/BBID/g
This commit aims to fix 2 issues. The first is that we (in #bookbrainz-devel)
think that identifiers should be unique to the context of the entire BookBrainz
application, not just a single concept. This commit moves in that direction, by
having a single list of identifiers. Note that this is *not* enough to enforce
the invariant we're looking for, but for all purposes it does the job now. It
will still allow you to have a book and an edition with the same ID in the
schema, but due to the way creating new concepts works now (always asking for a
new identifier), this probably won't happen in practice.

The second change in this commit is to stop refering to things as 'gid's and use
the more type agnostic term BBID. Kuno/warp mentioned that there's no real need
for GIDs here, and he has a point; BookBrainz is the only source capable of
creating generating IDs and can guarantee their uniqueness, so UUIDs aren't
entirely necessary. It might be possible to use something more human-friendly. I
don't know if we *will* make this change, but the possibility is a bit more
there now.