Move a few things around
All the data type definitions are now in the same file. I don't see much benefit
of having each one in its own file (I guess I'm too used to needing that in
other projects!). I've moved a few of the conversion instances out of the type
declarations too. I want to make all of this as independent of storage as
possible, mostly.