Add the basics of authentication
This commit combines all my prior work on an authentication system (a
history-rewrite of the auth branch). Currently this supports:

* Registration, with basic validation (non-empty fields, passwords must match)
* Login, with basic validation (non-empty fields)
* A "remember me" token
* Login redirection, for where to go once login is successful
* Logging out

This is currently implemented with a custom authentication backend for Snap,
though in the future this will almost certainly change to use the HDBC Snaplet,
when that is more mature.

None of the actions are yet exposed from the website, users have to manually
enter URLs.