Allow adding editions with country, language and format
These are the first fields that are populated with values from the database. The
edition form is now generated in a monadic context, which means it can have
access to the database to decide which values are possible. I've introduced a
very basic field generater, optionalDbSelect, which creates a <select> element
with an optional null value at the top of the list.