bookbrainz-search: Do not use PostgresSQL notifications
I spoke with the PgQ author and he is not too eager to implement this. I don't
have any good arguments as to why it's a good idea, so I will trust his
judgement that 'sleep' is the way to go. I imagine that I will move to RabbitMQ
before this becomes a problem anyway (as we need it for other time critical
stuff, like cache invalidation).